Completely change the underlying data structure. Closes #33.
Operating on WAI requests was problematic, as you can only access the
parameters and files once, as they are stored in an IO Source. I "solved"
this in Λδ by reading everything once and passing it around, but that
made dealing with forms a pain. In adding nice accessors for parameters
I fixed this issue:

 - Add nice accessors for params/files
 - Change from using Request to using Cry, a Seacat request, which has
   a reference to the params and files